Moved GraphQL options into /graphql/index.ts
