+++ src/server/app.ts
@@ -1,7 +1,7 @@
 import express from 'express'
 import graphqlHttp from 'express-graphql'
-import { buildSchema } from 'graphql'
 
+import graphqlOptions from '../graphql'
 import { Exercise } from '../database/models/Exercise'
 import { MuscleGroup } from '../database/models/MuscleGroup'
 import { Equipment } from '../database/models/Equipment'
@@ -10,52 +10,7 @@ const app = express()
 
 app.use(express.json())
 
-app.use(
-    '/graphql',
-    graphqlHttp({
-        graphiql: true,
-        schema: buildSchema(`
-        type Exercise {
-            id: ID!
-            name: String!
-            level: Int!
-            rating: Int
-            createdAt: String
-            updatedAt: String
-        }
-
-        type MuscleGroup {
-            id: ID!
-            name: String!
-            exercises: [Exercise!]!
-            createdAt: String
-            updatedAt: String
-        }
-
-        input MuscleGroupInput {
-            name: String!
-        }
-
-        type RootQuery {
-            muscleGroups: [MuscleGroup!]!
-        }
-
-        schema {
-            query: RootQuery
-        }
-        `),
-        rootValue: {
-            muscleGroups: async () => {
-                try {
-                    const muscleGroups = await MuscleGroup.findAll()
-                    return muscleGroups
-                } catch (err) {
-                    throw err
-                }
-            }
-        }
-    })
-)
+app.use('/graphql', graphqlHttp(graphqlOptions))
 
 app.get('/', (req: express.Request, res: express.Response, next) => {
     res.send('hello')
